7. admin 패키지명 변경 및 user 수정.
