--- fullcourse/src/main/java/com/example/fullcourse/admin/festival/AdminFestivalController.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.festival;
-
-public class AdminFestivalController {
-
-}
