--- fullcourse/src/main/java/com/example/fullcourse/admin/festival/mapper/AdminFestivalMapper.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.festival.mapper;
-
-public interface AdminFestivalMapper {
-
-}
