--- fullcourse/src/main/java/com/example/fullcourse/admin/notice/AdminNoticeController.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.notice;
-
-public class AdminNoticeController {
-
-}
