--- fullcourse/src/main/java/com/example/fullcourse/admin/notice/mapper/AdminNoticeMapper.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.notice.mapper;
-
-public interface AdminNoticeMapper {
-
-}
