--- fullcourse/src/main/java/com/example/fullcourse/admin/product/AdminProductController.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.product;
-
-public class AdminProductController {
-
-}
