--- fullcourse/src/main/java/com/example/fullcourse/admin/product/mapper/AdminProductMapper.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.product.mapper;
-
-public interface AdminProductMapper {
-
-}
