--- fullcourse/src/main/java/com/example/fullcourse/admin/report/AdminReportController.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.report;
-
-public class AdminReportController {
-
-}
