--- fullcourse/src/main/java/com/example/fullcourse/admin/report/mapper/AdminReportMapper.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.report.mapper;
-
-public interface AdminReportMapper {
-
-}
