--- fullcourse/src/main/java/com/example/fullcourse/admin/tour/AdminTourController.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.tour;
-
-public class AdminTourController {
-
-}
