--- fullcourse/src/main/java/com/example/fullcourse/admin/tour/mapper/AdminTourMapper.java
@@ -1,5 +1,0 @@
-package com.example.fullcourse.admin.tour.mapper;
-
-public interface AdminTourMapper {
-
-}
