+++ fullcourse/src/main/java/com/fullcourse/admin/festival/AdminFestivalController.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.festival;
+
+public class AdminFestivalController {
+
+}
