+++ fullcourse/src/main/java/com/fullcourse/admin/festival/mapper/AdminFestivalMapper.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.festival.mapper;
+
+public interface AdminFestivalMapper {
+
+}
