+++ fullcourse/src/main/java/com/fullcourse/admin/notice/AdminNoticeController.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.notice;
+
+public class AdminNoticeController {
+
+}
