+++ fullcourse/src/main/java/com/fullcourse/admin/notice/mapper/AdminNoticeMapper.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.notice.mapper;
+
+public interface AdminNoticeMapper {
+
+}
