+++ fullcourse/src/main/java/com/fullcourse/admin/product/AdminProductController.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.product;
+
+public class AdminProductController {
+
+}
