+++ fullcourse/src/main/java/com/fullcourse/admin/product/mapper/AdminProductMapper.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.product.mapper;
+
+public interface AdminProductMapper {
+
+}
