+++ fullcourse/src/main/java/com/fullcourse/admin/report/AdminReportController.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.report;
+
+public class AdminReportController {
+
+}
