+++ fullcourse/src/main/java/com/fullcourse/admin/report/mapper/AdminReportMapper.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.report.mapper;
+
+public interface AdminReportMapper {
+
+}
