+++ fullcourse/src/main/java/com/fullcourse/admin/tour/AdminTourController.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.tour;
+
+public class AdminTourController {
+
+}
