+++ fullcourse/src/main/java/com/fullcourse/admin/tour/mapper/AdminTourMapper.java
@@ -1,0 +1,5 @@
+package com.fullcourse.admin.tour.mapper;
+
+public interface AdminTourMapper {
+
+}
