+++ fullcourse/src/main/java/com/fullcourse/admin/user/mapper/AdminUserMapper.java
@@ -1,11 +1,11 @@
-package com.example.fullcourse.admin.user.mapper;
+package com.fullcourse.admin.user.mapper;
 
 import java.util.List;
 import java.util.Map;
 
 import org.apache.ibatis.annotations.Mapper;
 
-import com.example.fullcourse.admin.user.UserVO;
+import com.fullcourse.admin.user.UserVO;
 
 @Mapper
 public interface AdminUserMapper {
